Split out the input handling
